requirements.txt
Pillow>=9.5.0
streamlit==1.25.0
streamlit_drawable_canvas==0.9.3
sympy==1.13.3
numpy==1.24.0